requirements.txt
Babel==2.5.3
certifi==2018.4.16
cloudinary==1.11.0
confusable-homoglyphs==3.0.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap4==0.0.6
django-phonenumber-field==2.0.0
django-registration==2.4.1
djangorestframework==3.8.2
gunicorn==19.7.1
mock==2.0.0
pbr==4.0.2
phonenumberslite==8.9.4
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-decouple==3.1
pytz==2018.3
six==1.11.0
urllib3==1.22
whitenoise==3.3.1
